DNS callback to use new xDNSDoCallback API (#985)
* fix issue with DNS callback

* update test cli code

* fix formatting

* fix build warnings

* fix formatting

* clang build fix

+++ source/FreeRTOS_DNS_Callback.c
@@ -151,7 +151,7 @@
             if( listLIST_IS_EMPTY( &xCallbackList ) != pdFALSE )
             {
                 /* This is the first one, start the DNS timer to check for timeouts */
-                vDNSTimerReload( FreeRTOS_min_uint32( 1000U, uxTimeout ) );
+                vDNSTimerReload( FreeRTOS_min_uint32( 1000U, ( uint32_t ) uxTimeout ) );
             }
 
             ( void ) strcpy( pxCallback->pcName, pcHostName );
